Update migrations and install model caching
Create base model that uses Cachable, all models (except User) extend base model now.
Add DepartmentFacultyUser model.
Add EventScope, EventType, UserGender and UserType enums.
Update User, StudentProfile, TeachingStaffProfile, CompanyProfile, ModeratorProfile and DepartmentFaculty relations.
Update User and Event accessors.
Update User, StudentProfile, TeachingStaffProfile, CompanyProfile, ModeratorProfile and Event factories.
Update user, student_profile, teaching_staff_profile, moderator_profile, events and rates migrations.
Update User and DepartmentFacultyUser seeders.

+++ database/migrations/2020_05_24_024959_create_rates_table.php
@@ -15,7 +15,7 @@ class CreateRatesTable extends Migration
 	{
 		Schema::create('rates', function (Blueprint $table) {
 			$table->id();
-			$table->bigInteger('rate');
+			$table->tinyInteger('rate');
 			$table->unsignedBigInteger('comment_id');
 			$table->unsignedBigInteger('user_id');
 			$table->timestamps();
